new transform api works, sample_out left
