+++ hub/core/transform/__init__.py
@@ -1,1 +1,1 @@
-from hub.core.transform.transform import run_pipeline, store_shard, transform
+from hub.core.transform.transform import parallel, compose
