+++ hub/core/transform/test_transform.py
@@ -1,14 +1,15 @@
+from hub.core.transform.transform import Pipeline
 from hub.util.exceptions import MemoryDatasetNotSupportedError
 import pytest
 from hub.core.storage.memory import MemoryProvider
 from hub.util.remove_cache import remove_memory_cache
-from hub import transform  # type: ignore
 import numpy as np
 import hub
 from hub.tests.dataset_fixtures import enabled_datasets
 from click.testing import CliRunner
 
 
+@hub.parallel
 def fn1(i, mul=1, copy=1):
     d = {}
     d["image"] = np.ones((337, 200)) * i * mul
@@ -16,11 +17,13 @@ def fn1(i, mul=1, copy=1):
     return d if copy == 1 else [d] * copy
 
 
+@hub.parallel
 def fn2(sample, mul=1, copy=1):
     d = {"image": sample["image"] * mul, "label": sample["label"] * mul}
     return d if copy == 1 else [d] * copy
 
 
+@hub.parallel
 def fn3(i, mul=1, copy=1):
     d = {}
     d["image"] = np.ones((1310, 2087)) * i * mul
@@ -41,9 +44,8 @@ def test_single_transform_hub_dataset(ds):
         ds_out = ds
         ds_out.create_tensor("image")
         ds_out.create_tensor("label")
-        transform(
-            data_in, fn2, ds_out, pipeline_kwargs={"copy": 1, "mul": 2}, workers=5
-        )
+
+        fn2(copy=1, mul=2).eval(data_in, ds_out, workers=5)
         assert len(ds_out) == 99
         for index in range(1, 100):
             np.testing.assert_array_equal(
@@ -70,9 +72,7 @@ def test_single_transform_hub_dataset_htypes(ds):
         ds_out = ds
         ds_out.create_tensor("image")
         ds_out.create_tensor("label")
-        transform(
-            data_in, fn2, ds_out, pipeline_kwargs={"copy": 1, "mul": 2}, workers=5
-        )
+        fn2(copy=1, mul=2).eval(data_in, ds_out, workers=5)
         assert len(ds_out) == 99
         for index in range(1, 100):
             np.testing.assert_array_equal(
@@ -92,13 +92,8 @@ def test_chain_transform_list_small(ds):
     ds_out = ds
     ds_out.create_tensor("image")
     ds_out.create_tensor("label")
-    transform(
-        ls,
-        [fn1, fn2],
-        ds_out,
-        workers=1,
-        pipeline_kwargs=[{"mul": 5, "copy": 2}, {"mul": 3, "copy": 3}],
-    )
+    pipeline = hub.compose([fn1(mul=5, copy=2), fn2(mul=3, copy=3)])
+    pipeline.eval(ls, ds_out, workers=5)
     assert len(ds_out) == 600
     for i in range(100):
         for index in range(6 * i, 6 * i + 6):
@@ -117,13 +112,8 @@ def test_chain_transform_list_big(ds):
     ds_out = ds
     ds_out.create_tensor("image")
     ds_out.create_tensor("label")
-    transform(
-        ls,
-        [fn3, fn2],
-        ds_out,
-        workers=3,
-        pipeline_kwargs=[{"mul": 5, "copy": 2}, {"mul": 3, "copy": 2}],
-    )
+    pipeline = hub.compose([fn3(mul=5, copy=2), fn2(mul=3, copy=3)])
+    pipeline.eval(ls, ds_out, workers=3)
     assert len(ds_out) == 8
     for i in range(2):
         for index in range(4 * i, 4 * i + 4):
@@ -143,17 +133,11 @@ def test_chain_transform_list_small_processed(ds):
     ds_out.create_tensor("label")
     if isinstance(remove_memory_cache(ds.storage), MemoryProvider):
         with pytest.raises(MemoryDatasetNotSupportedError):
-            transform(ls, fn1, ds_out, scheduler="processed")
+            fn2().eval(ls, ds_out, workers=3, scheduler="processed")
         return
 
-    transform(
-        ls,
-        [fn1, fn2],
-        ds_out,
-        workers=3,
-        pipeline_kwargs=[{"mul": 5, "copy": 2}, {"mul": 3, "copy": 3}],
-        scheduler="processed",
-    )
+    pipeline = hub.compose([fn1(mul=5, copy=2), fn2(mul=3, copy=3)])
+    pipeline.eval(ls, ds_out, workers=3, scheduler="processed")
     assert len(ds_out) == 600
     for i in range(100):
         for index in range(6 * i, 6 * i + 6):
